remove pre-rendering of the markdown
