+++ cmd/slackdump/internal/workspace/list.go
@@ -20,11 +20,11 @@ import (
 var CmdWspList = &base.Command{
 	UsageLine: "slackdump workspace list [flags]",
 	Short:     "list saved workspaces",
-	Long: base.Render(`
+	Long: `
 # Workspace List Command
 
 **List** allows to list Slack Workspaces, that you have previously authenticated in.
-`),
+`,
 	FlagMask:   flagmask,
 	PrintFlags: true,
 }
